+++ src/test/01_EventTypeTests.ts
@@ -3,7 +3,7 @@ import {EventType} from "./../EventType";
 
 let expect = chai.expect;
 
-describe("EventType tests", () => {
+describe("EventType", () => {
 	it("ConnectorAdded should result in true when compared to its string representation", (done) => {
 		expect(EventType.ConnectorAdded).to.deep.equal("ConnectorAddedEvent");
 		done();